backend/pyproject.toml
[tool.poetry]
name = "database-dashboard-backend"
version = "0.1.0"
description = "Backend for Database Dashboard using FastAPI"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
uvicorn = "^0.17.0"
sqlalchemy = "^1.4.27"
alembic = "^1.7.5"
pydantic = "^1.9.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
httpx = "^0.19.0"
mypy = "^0.910"
black = "^21.12b0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"